+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to
 
 ### Added
 
+- ✨(domains) allow to re-run check on domain if status is failed
 - ✨(organization) add `is_active` field
 - ✨(domains) notify support when domain status changes #668
 - ✨(domains) define domain check interval as a settings